new strata run so north-south splits are right
The week of May 19 we discovered that the n/s split from the VAST output was nearly 50/50, which seemed odd. KFJ tracked this down to a known code error which mislabeled the strata, so we re-ran VAST in the NWFSCASSESS package and found overall estimates <100mt different from the run done by LQ earlier this year. for time's sake we stuck with the same overall index used in the base model, but are showing in plots the values here.